+++ policyengine_us/variables/gov/states/ny/tax/income/credits/ctc/ny_ctc.py
@@ -22,17 +22,17 @@ class ny_ctc(Variable):
         age = person("age", period)
         if not p.pre_tcja:
             federal_ctc = tax_unit("ctc", period)
+            gov = parameters(period).gov
+            qualifies_for_federal_ctc = person("ctc_qualifying_child", period)
         else:
-            # Initialise pre-TCJA CTC branch
-
+            # Initialise pre-TCJA CTC branch and parameters.
             simulation = tax_unit.simulation
             pre_tcja_ctc = simulation.get_branch("pre_tcja_ctc")
             pre_tcja_ctc.tax_benefit_system = (
                 simulation.tax_benefit_system.clone()
             )
-
             branch_parameters = pre_tcja_ctc.tax_benefit_system.parameters
-
+            # Update parameters to pre-TCJA values.
             for (
                 ctc_parameter
             ) in branch_parameters.gov.irs.credits.ctc.get_descendants():
@@ -42,11 +42,11 @@ class ny_ctc(Variable):
                         stop=instant("2026-01-01"),
                         value=ctc_parameter("2017-01-01"),
                     )
-
+            # Delete all arrays from pre-TCJA CTC branch.
             for variable in pre_tcja_ctc.tax_benefit_system.variables:
                 if "ctc" in variable:
                     pre_tcja_ctc.delete_arrays(variable)
-
+            # Calculate pre-TCJA CTC.
             maximum_ctc = pre_tcja_ctc.calculate(
                 "ctc_child_individual_maximum", period
             )
@@ -57,8 +57,11 @@ class ny_ctc(Variable):
                 maximum_ctc * meets_ny_minimum_age,
             )
             federal_ctc = pre_tcja_ctc.tax_unit("ctc_value", period)
-
-        qualifies_for_federal_ctc = person("ctc_qualifying_child", period)
+            qualifies_for_federal_ctc = pre_tcja_ctc.person(
+                "ctc_qualifying_child", period
+            )
+            gov = branch_parameters(period).gov
+        # Remaining logic is based on NY parameters.
         qualifies = qualifies_for_federal_ctc & (age >= p.minimum_age)
         qualifying_children = tax_unit.sum(qualifies)
         federal_match = federal_ctc * p.amount.percent
@@ -66,10 +69,9 @@ class ny_ctc(Variable):
         # minimum amount per child.
         minimum = p.amount.minimum * qualifying_children
         agi = tax_unit("adjusted_gross_income", period)
-        # Uses pre-TCJA parameters.
-        pre_tcja_ctc = parameters("2017-01-01").gov.irs.credits.ctc
-        filing_status = tax_unit("filing_status", period)
-        federal_threshold = pre_tcja_ctc.phase_out.threshold[filing_status]
+        federal_threshold = gov.irs.credits.ctc.phase_out.threshold[
+            tax_unit("filing_status", period)
+        ]
         eligible_for_minimum = agi < federal_threshold
         applicable_minimum = eligible_for_minimum * minimum
         eligible = qualifying_children > 0